requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
altair==5.2.0
annotated-types==0.6.0
anyio==4.3.0
asgiref==3.8.1
async-timeout==4.0.3
attrs==23.2.0
backoff==2.2.1
bcrypt==4.1.2
blinker==1.7.0
blis==0.7.11
build==1.2.1
cachetools==5.3.3
catalogue==2.0.10
certifi==2024.2.2
charset-normalizer==3.3.2
chroma-hnswlib==0.7.3
chromadb==0.4.24
click==8.1.7
cloudpathlib==0.16.0
colorama==0.4.6
coloredlogs==15.0.1
comtypes==1.3.1
confection==0.1.4
cymem==2.0.8
Cython==3.0.9
dataclasses==0.6
dataclasses-json==0.6.4
Deprecated==1.2.14
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1-py3-none-any.whl#sha256=86cc141f63942d4b2c5fcee06630fd6f904788d2f0ab005cce45aadb8fb73889
exceptiongroup==1.2.0
faiss-cpu==1.8.0
fastapi==0.110.1
filelock==3.13.4
Flask==3.0.2
Flask-Cors==4.0.0
flatbuffers==24.3.25
frozenlist==1.4.1
fsspec==2024.3.1
gitdb==4.0.11
GitPython==3.1.42
google-ai-generativelanguage==0.4.0
google-api-core==2.18.0
google-auth==2.29.0
google-generativeai==0.4.1
googleapis-common-protos==1.63.0
greenlet==3.0.3
grpcio==1.62.1
grpcio-status==1.62.1
h11==0.14.0
httptools==0.6.1
huggingface-hub==0.22.2
humanfriendly==10.0
idna==3.6
importlib-metadata==7.0.0
importlib_resources==6.4.0
itsdangerous==2.1.2
Jinja2==3.1.3
joblib==1.3.2
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
kubernetes==29.0.0
langchain==0.1.13
langchain-community==0.0.31
langchain-core==0.1.40
langchain-google-genai==0.0.11
langchain-text-splitters==0.0.1
langcodes==3.3.0
langsmith==0.1.33
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.21.1
mdurl==0.1.2
mmh3==4.1.0
monotonic==1.6
mpmath==1.3.0
multidict==6.0.5
murmurhash==1.0.10
mypy-extensions==1.0.0
nameparser==1.1.3
nltk==3.8.1
numpy==1.26.4
oauthlib==3.2.2
onnxruntime==1.17.1
opentelemetry-api==1.24.0
opentelemetry-exporter-otlp-proto-common==1.24.0
opentelemetry-exporter-otlp-proto-grpc==1.24.0
opentelemetry-instrumentation==0.45b0
opentelemetry-instrumentation-asgi==0.45b0
opentelemetry-instrumentation-fastapi==0.45b0
opentelemetry-proto==1.24.0
opentelemetry-sdk==1.24.0
opentelemetry-semantic-conventions==0.45b0
opentelemetry-util-http==0.45b0
orjson==3.9.15
overrides==7.7.0
packaging==23.2
pandas==2.2.1
pillow==10.2.0
posthog==3.5.0
preshed==3.0.9
proto-plus==1.23.0
protobuf==4.25.3
pulsar-client==3.4.0
pyarrow==15.0.2
pyasn1==0.6.0
pyasn1_modules==0.4.0
PyAudio==0.2.14
pydantic==2.6.4
pydantic_core==2.16.3
pydeck==0.8.1b0
Pygments==2.17.2
PyPDF2==3.0.1
PyPika==0.48.9
pypiwin32==223
pyproject_hooks==1.0.0
pyreadline3==3.4.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pyttsx3==2.90
pytz==2024.1
pywin32==306
PyYAML==6.0.1
referencing==0.34.0
regex==2023.12.25
requests==2.31.0
requests-oauthlib==2.0.0
rich==13.7.1
rpds-py==0.18.0
rsa==4.9
scipy==1.12.0
six==1.16.0
smart-open==6.4.0
smmap==5.0.1
sniffio==1.3.1
spacy==3.7.4
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SpeechRecognition==3.10.3
SQLAlchemy==2.0.29
srsly==2.4.8
starlette==0.37.2
streamlit==1.32.2
sympy==1.12
tenacity==8.2.3
thinc==8.2.3
tokenizers==0.15.2
toml==0.10.2
tomli==2.0.1
toolz==0.12.1
tornado==6.4
tqdm==4.66.2
typer==0.9.4
typing-inspect==0.9.0
typing_extensions==4.10.0
tzdata==2024.1
urllib3==2.2.1
uvicorn==0.29.0
wasabi==1.1.2
watchdog==4.0.0
watchfiles==0.21.0
weasel==0.3.4
websocket-client==1.7.0
websockets==12.0
Werkzeug==3.0.2
wrapt==1.16.0
yarl==1.9.4
zipp==3.18.1
